requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
psycopg2-binary
python-decouple
sqlparse
tzdata
requests
